backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
openai>=1.30.0
langchain>=0.2.1
langchain-openai>=0.1.8
langchain-community>=0.2.1
qdrant-client==1.9.1
pypdf==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4